requirements.txt
Flask
Flask-SQLAlchemy
langchain
langgraph
python-dotenv
openai
deepgram-sdk>=3.0
pydub
Markdown
requests
